the task #8

+++ src/main/java/Task5/operation/Addition.java
@@ -1,7 +1,7 @@
 package Task5.operation;
 
 public class Addition extends Operation {
-    public float getResult(float a, float b){
+    public double getResult(double a, double b){
         return a+b;
     }
 }
