+++ src/main/java/Task5/operation/Subtraction.java
@@ -1,7 +1,7 @@
 package Task5.operation;
 
 public class Subtraction extends Operation {
-    public float getResult(float a, float b){
+    public double getResult(double a, double b){
         return a-b;
     }
 }
